Refactored Generic HipChat bot to be a class, cleaned up jsdoc and jscs errors
